Refactoring
TicTacToe: added player; changed no-param defs to lazy vals;
Shuffle: added seed parameter.
TreeSpec: tagged one test as Slow.
Evaluator: moved into moves package.
