+++ src/main/scala/com/phasmidsoftware/util/Shuffle.scala
@@ -6,12 +6,11 @@ import scala.collection.mutable.ListBuffer
 
 object Shuffle {
 
-  def apply[X](xs: Seq[X]): Seq[X] = {
+  def apply[X](xs: Iterable[X], seed: Long = System.currentTimeMillis()): Seq[X] = {
     import scala.jdk.CollectionConverters._
     val listBuffer: ListBuffer[X] = new ListBuffer().appendAll(xs)
     val jList: util.List[X] = listBuffer.asJava
-    java.util.Collections.shuffle(jList, new Random(3L))
+    java.util.Collections.shuffle(jList, new Random(seed))
     jList.asScala.toSeq
   }
-
 }
